Game.tsxのロジックやUIの切り分け
