+++ src/types/types.ts
@@ -1,0 +1,6 @@
+type Square = {
+  id: number;
+  state: "○" | "×" | "";
+};
+
+export type { Square };
