+++ src/utils/IsValid.ts
@@ -1,0 +1,7 @@
+const IsValid = (text: string): boolean => {
+  const regex =
+    /^[\u3040-\u309F\u30A0-\u30FF\u4E00-\u9FFFa-zA-Z0-9\uFF10-\uFF19\uFF21-\uFF3A\uFF41-\uFF5A]*$/;
+  return regex.test(text);
+};
+
+export default IsValid;
